backend/requirements.txt
Flask==2.3.4
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
PyJWT==2.8.0
psycopg2-binary==2.9.7
Flask-Cors==4.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1